Move Question.alertType to utils/notifications as getAlertType
