--- frontend/src/metabase-lib/v1/Alert/constants.ts
@@ -1,12 +1,0 @@
-export const ALERT_TYPE_ROWS = "alert-type-rows";
-export const ALERT_TYPE_TIMESERIES_GOAL = "alert-type-timeseries-goal";
-export const ALERT_TYPE_PROGRESS_BAR_GOAL = "alert-type-progress-bar-goal";
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars -- used for types
-const AlertTypes = [
-  ALERT_TYPE_ROWS,
-  ALERT_TYPE_TIMESERIES_GOAL,
-  ALERT_TYPE_PROGRESS_BAR_GOAL,
-] as const;
-
-export type NotificationTriggerType = (typeof AlertTypes)[number];
